feat(api): add category, month reporting
Users may allocate of (or portions of) any DEPOSIT transaction directly to a category, counting it as part of that category's balance, OR they may now choose for such a transaction to not contribute to a category directly whatsoever.

Users can now directly assign money to categories using the assignment endpoints, and can get reports on any of one or all categories as they please.