refactor(tests): move req caller into api_test
The pinchertest package (as it is currently titled) could actually be potentially useful as its own package to be used by client applications. Even without that external use in mind, the use of httptest is the only thing REALLY making the pinchertest package have anything to do with testing itself. So Call() ought to just be a helper function within the api_test file.