refactor(api): adapt Test_MakeLoginDeleteUsers to case-loop implementation
Well...I was excited about the prospect of an entirely case-loop approach to the httptest request-mocking integration test implementation. But while possible, it seems to me that up to a point, it is an overengineered approach that can make a resource-heavy test LESS readable, as opposed to MORE readable, as intended.

This httpTestCase may yet be useful, so I'll keep it around for now.