feat(auth)!: switch encryption algorithm from bcrypt to argon2id
Note that in any scenario wherein this application released (and
therefore KNOWN to be running in production), it ought be
best practice to rehash passwords with argon2 as logins came in, rather
than altogether barring existing users from accessing accounts whose
passwords were hashed with bcrypt. Then this breaking change could come
further down the road, instead of pushing an update so abruptly.